requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
alembic==1.11.1
python-dotenv==1.0.0
python-keycloak==3.0.0
PyJWT==2.8.0
itsdangerous==2.1.2
APScheduler==3.10.1
pydantic==1.10.11
jinja2==3.1.2
email-validator==2.1.0
requests==2.31.0
pytest==7.4.2
